Orders create and fetch services

+++ models/product.model.js
@@ -21,7 +21,7 @@ module.exports = (sequelizeInstance) => {
             type: DataTypes.TEXT
         },
         price: {
-            type: DataTypes.DECIMAL(10, 2),
+            type: DataTypes.DOUBLE(10, 2),
             allowNull: false,
             default: 0
         }
